% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.dataset.R
\name{chicken_chip}
\alias{chicken_chip}
\title{chicken chip}
\source{
Groenen et al
}
\usage{
chicken_chip
}
\description{
Genome for chicken according to Groenen et al.
}
\author{
Torsten Pook \email{torsten.pook@uni-goettingen.de}
}
