% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.bve.R
\name{get.bve}
\alias{get.bve}
\title{Export estimated breeding values}
\usage{
get.bve(population, database = NULL, gen = NULL, cohorts = NULL)
}
\arguments{
\item{population}{Population list}

\item{database}{Groups of individuals to consider for the export}

\item{gen}{Quick-insert for database (vector of all generations to export)}

\item{cohorts}{Quick-insert for database (vector of names of cohorts to export)}
}
\description{
Function to export estimated breeding values
}
